+++ tools/ssl_utils/reliable_student.py
@@ -1,63 +1,43 @@
 import torch
-from .semi_utils import reverse_transform, load_data_to_gpu, construct_pseudo_label
+from .semi_utils import reverse_transform, load_data_to_gpu, construct_pseudo_label, construct_pseudo_label_scores
 from pcdet.models.model_utils.model_nms_utils import class_agnostic_nms
 from pcdet.ops.iou3d_nms import iou3d_nms_utils
 import os
 import pickle
 
-@torch.no_grad()
-def construct_pseudo_label_scores(boxes):
-    score_list = []
-    num_gt_list = []
-    for box in boxes:
-        pred_scores = box['pred_scores']
-        num_gt_list.append(pred_scores.shape[0])
-        score_list.append(pred_scores)
-    batch_size = len(boxes)
-    num_max_gt = max(num_gt_list)
-    gt_boxes = score_list[0].new_zeros((batch_size, num_max_gt))
-    for bs_idx in range(batch_size):
-        num_gt = num_gt_list[bs_idx]
-        gt_boxes[bs_idx, :num_gt] = score_list[bs_idx]
-    return gt_boxes
+
 
 @torch.no_grad()
 def filter_pseudo_labels(pred_dicts, cfgs):
     filtered_pred_dicts = []
     for ind in range(len(pred_dicts)):
+        record_dict = {
+            'pred_boxes': torch.empty((0, 7)),
+            'pred_scores': torch.empty((0, 1)),
+            'pred_labels': torch.empty((0, 1)),
+            'pred_sem_scores': torch.empty((0, 1))
+            }
         pseudo_score = pred_dicts[ind]['pred_scores']
         pseudo_box = pred_dicts[ind]['pred_boxes']
         pseudo_label = pred_dicts[ind]['pred_labels']
         pseudo_sem_score = pred_dicts[ind]['pred_sem_scores']
-        if len(pseudo_label) == 0:
-            record_dict = {
-                'pred_boxes': torch.empty((0, 7)),
-                'pred_scores': torch.empty((0, 1)),
-                'pred_labels': torch.empty((0, 1)),
-                'pred_sem_scores': torch.empty((0, 1))
-            }
-            filtered_pred_dicts.append(record_dict)
-            continue
-
-        conf_thresh = torch.tensor(cfgs.TEACHER.THRESH, device=pseudo_label.device).unsqueeze(
-            0).repeat(len(pseudo_label), 1).gather(dim=1, index=(pseudo_label - 1).unsqueeze(-1))
-
-        sem_conf_thresh = torch.tensor(cfgs.TEACHER.SEM_THRESH, device=pseudo_label.device).unsqueeze(
-            0).repeat(len(pseudo_label), 1).gather(dim=1, index=(pseudo_label - 1).unsqueeze(-1))
-
-        valid_inds = pseudo_score > conf_thresh.squeeze()
-
-        valid_inds = valid_inds & (pseudo_sem_score > sem_conf_thresh.squeeze())
-
-        record_dict = {
-            'pred_boxes': pseudo_box[valid_inds],
-            'pred_scores': pseudo_score[valid_inds],
-            'pred_labels': pseudo_label[valid_inds],
-            'pred_sem_scores': pseudo_sem_score[valid_inds]
-        }
+        if len(pseudo_label) > 0:
+            conf_thresh = torch.tensor(cfgs.TEACHER.THRESH, device=pseudo_label.device).unsqueeze(
+                0).repeat(len(pseudo_label), 1).gather(dim=1, index=(pseudo_label - 1).unsqueeze(-1))
+            sem_conf_thresh = torch.tensor(cfgs.TEACHER.SEM_THRESH, device=pseudo_label.device).unsqueeze(
+                0).repeat(len(pseudo_label), 1).gather(dim=1, index=(pseudo_label - 1).unsqueeze(-1))
+            valid_inds = pseudo_score > conf_thresh.squeeze()
+            valid_inds = valid_inds & (pseudo_sem_score > sem_conf_thresh.squeeze())
+            if valid_inds.sum():
+                record_dict = {
+                    'pred_boxes': pseudo_box[valid_inds],
+                    'pred_scores': pseudo_score[valid_inds],
+                    'pred_labels': pseudo_label[valid_inds],
+                    'pred_sem_scores': pseudo_sem_score[valid_inds]
+                }
         filtered_pred_dicts.append(record_dict)
 
-    return pred_dicts
+    return filtered_pred_dicts
 
 
 @torch.no_grad()
@@ -91,8 +71,8 @@ def reliable_student(teacher_model, student_model,
     load_data_to_gpu(ud_student_batch_dict)
     load_data_to_gpu(ud_teacher_batch_dict)
     if dist:
-        teacher_model=teacher_model.module.onepass
-        student_model=student_model.module.onepass
+        teacher_model=teacher_model.onepass
+        student_model=student_model.onepass
 
     with torch.no_grad():
         for cur_module in teacher_model.module_list:
